BCF-2165:  mv cli config validate command (#8807)
