+++ core/main_test.go
@@ -164,7 +164,6 @@ func ExampleRun_config() {
 	//    show      Show the application configuration
 	//    loglevel  Set log level
 	//    logsql    Enable/disable SQL statement logging
-	//    validate  Validate provided TOML config file, and print the full effective configuration, with defaults included
 	//
 	// OPTIONS:
 	//    --help, -h  show help
@@ -429,6 +428,7 @@ func ExampleRun_node() {
 	//    rebroadcast-transactions  Manually rebroadcast txs matching nonce range with the specified gas price. This is useful in emergencies e.g. high gas prices and/or network congestion to forcibly clear out the pending TX queue
 	//    status                    Displays the health of various services running inside the node.
 	//    profile                   Collects profile metrics from the node.
+	//    validate                  Validate the TOML configuration and secrets that are passed as flags to the `node` command. Prints the full effective configuration, with defaults included
 	//    db                        Commands for managing the database.
 	//
 	// OPTIONS:
